Add some example config options.
Expose the type of configuration (this might be a bad choice, perhaps less directly expose the Debian choice, and maybe be more opinionated?).
Expose the list of other hostnames that should be considered local - this definitely should be a configuration option.

Use Debian's update-exim4.conf tool to handle creating the configuration file for now. If we end up managing the configuration in more detail in the future, then we'd probably want to do this ourselves.