Pebble should handle the 'daemonising', not Exim itself.
Note that this is not currently working. I think it's because Exim uses the -bs option in two ways, one to get SMTP input with stdin/stdout, and one where it listens but doesn't daemonise. It decides which case is being used by checking if stdin is a TCP/IP socket. However, I don't know if there's any way for Pebble to adjust that, or whether Pebble is really expecting to use stdin itself. Maybe it can be done by not starting immediately, then changing stdin, and then starting? Maybe Exim just isn't Pebble-friendly and would need to be patched for this situation.