Run cron to manage the queue.
We don't want to have Exim daemonise itself (with '-q <internal>'), but have pebble manage the
service entirely (both queue running and listening for connections. That means that we need to have
something that will periodically do a queue run. cron is already installed on the image, so use that
with a single task crontab. The cron daemon is not running (and we want all services managed by
Pebble) so add a pebble service for cron as well.