Simplify the exim config logic.
Use a base template ourselves and string formatting (to be done) rather than using the update-exim4.conf tool, which was getting pretty messy with all of the configuration.

Also hook up running the table creation from a SQL file, although this is not active until mysqlclient works (or a different MySQL client library is used).