Fix building mysqlclient.
There's an (not well documented but that will soon be fixed) option to specify packages that are required for the build, so use that to ensure that mysqlclient can build even though there is no Linux wheel on PyPI.

Also leave two other commented methods of handling this issue for documentation/learning purposes.

Note that MySQLdb still cannot be imported, because there's a package that needs to be installed on the sidecar container that has the agent. To be fixed soon, I hope.